+++ wfx_fmac_driver/secure_link/sl_wfx_secure_link.c
@@ -212,6 +212,11 @@ sl_status_t sl_wfx_secure_link_renegotiate_session_key(void)
   result = sl_wfx_get_status_code(sl_wfx_htole32(exchange_pub_keys_ind->body.status), SL_WFX_SECURELINK_EXCHANGE_PUB_KEYS_IND_ID);
   SL_WFX_ERROR_CHECK(result);
 
+  // reset nonce counters as soon as possible
+  sl_wfx_context->secure_link_nonce.hp_packet_count = 0;
+  sl_wfx_context->secure_link_nonce.rx_packet_count = 0;
+  sl_wfx_context->secure_link_nonce.tx_packet_count = 0;
+
   result = sl_wfx_host_verify_pub_key(exchange_pub_keys_ind, sl_wfx_context->secure_link_mac_key, sl_host_pub_key);
   SL_WFX_ERROR_CHECK(result);
 
@@ -242,7 +247,6 @@ sl_status_t sl_wfx_secure_link_renegotiate_session_key(void)
 sl_status_t sl_wfx_secure_link_configure(const uint8_t *encryption_bitmap, uint8_t disable_session_key_protection)
 {
   sl_status_t result;
-  uint8_t     i;
   sl_wfx_securelink_configure_req_body_t request;
   sl_wfx_securelink_configure_cnf_t *reply = NULL;
 
@@ -264,6 +268,149 @@ sl_status_t sl_wfx_secure_link_configure(const uint8_t *encryption_bitmap, uint8
     if (memcmp(request.encr_bmp, reply->body.encr_bmp, SL_WFX_SECURE_LINK_ENCRYPTION_BITMAP_SIZE)) {
       result = SL_STATUS_WIFI_WARNING;
     }
+  }
+
+  return result;
+}
+
+sl_status_t sl_wfx_secure_link_send(uint8_t command_id, sl_wfx_generic_message_t **request, uint16_t *request_length)
+{
+  sl_status_t result = SL_STATUS_OK;
+
+  if (sl_wfx_context->secure_link_renegotiation_state == SL_WFX_SECURELINK_RENEGOTIATION_PENDING
+      && command_id != SL_WFX_SECURELINK_EXCHANGE_PUB_KEYS_REQ_ID) {
+    result = SL_STATUS_FAIL;
+    goto error_handler;
+  }
+
+  if (sl_wfx_secure_link_encryption_required_get(command_id) == SL_WFX_SECURE_LINK_ENCRYPTION_REQUIRED) {
+    // Nonce for encryption should have RX and HP counters 0, only use TX counter
+    sl_wfx_nonce_t encryption_nonce = { 0, 0, sl_wfx_context->secure_link_nonce.tx_packet_count };
+
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+    sl_wfx_host_log("TX packet %lu\n", sl_wfx_context->secure_link_nonce.tx_packet_count);
+#endif
+
+    // Round up to next crypto block size the part that will be ciphered
+    *request_length = ((*request_length + 15 - 2) & ~15) + 2;
+
+    // Encrypt the data
+    result = sl_wfx_host_encode_secure_link_data(*request,
+                                                 *request_length - 2,
+                                                 sl_wfx_context->secure_link_session_key,
+                                                 (uint8_t *)&encryption_nonce);
+    SL_WFX_ERROR_CHECK(result);
+
+    // Write the secure link header
+    uint16_t *secure_link_header = (uint16_t *)((uint8_t *)*request - 4);
+    *secure_link_header = sl_wfx_htole16((uint16_t) (sl_wfx_context->secure_link_nonce.tx_packet_count & 0xFFFF));
+    secure_link_header++;
+    *secure_link_header = sl_wfx_htole16((uint16_t) (0x4000 | ( (sl_wfx_context->secure_link_nonce.tx_packet_count >> 16) & 0x3FFF)));
+
+    sl_wfx_context->secure_link_nonce.tx_packet_count++;
+
+    if (sl_wfx_context->secure_link_nonce.tx_packet_count > SL_WFX_SECURE_LINK_NONCE_WATERMARK
+        && sl_wfx_context->secure_link_renegotiation_state == SL_WFX_SECURELINK_DEFAULT) {
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+      sl_wfx_host_log("--SLK renegotiation needed--\r\n");
+#endif
+      //queue key re-negotiation
+      sl_wfx_context->secure_link_renegotiation_state = SL_WFX_SECURELINK_RENEGOTIATION_NEEDED;
+    }
+
+    // Add the secure link buffer overhead and transmit
+    *request_length += SL_WFX_SECURE_LINK_HEADER_SIZE + SL_WFX_SECURE_LINK_CCM_TAG_SIZE;
+    *request = (sl_wfx_generic_message_t *)((uint8_t *)*request - SL_WFX_SECURE_LINK_HEADER_SIZE);
+  }
+
+  error_handler:
+  return result;
+}
+
+sl_status_t sl_wfx_secure_link_receive(sl_wfx_generic_message_t **network_rx_buffer, uint32_t read_length)
+{
+  sl_status_t result = SL_STATUS_OK;
+  uint8_t has_encrypt_header;
+  // Bit 14/15 of second word indicates if the message is encrypted
+  uint8_t encrypt_type = ((*network_rx_buffer)->header.info & SL_WFX_MSG_INFO_SECURE_LINK_MASK) >> SL_WFX_MSG_INFO_SECURE_LINK_OFFSET;
+
+  // Currently only RX counter is expected
+  switch (encrypt_type) {
+    case 0x0: has_encrypt_header = SL_WFX_SECURE_LINK_ENCRYPTION_NOT_REQUIRED; break;
+    case 0x2: has_encrypt_header = SL_WFX_SECURE_LINK_ENCRYPTION_REQUIRED; break;
+    default:
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+      sl_wfx_host_log("--SLK wrong counter type--\r\n");
+#endif
+      result = SL_STATUS_FAIL;
+      SL_WFX_ERROR_CHECK(result);
+  }
+
+  if (has_encrypt_header) {
+    uint16_t *nonce_ptr = (uint16_t *) *network_rx_buffer;
+    uint32_t new_packet_count = sl_wfx_unpack_16bit_little_endian(&((*network_rx_buffer)->header.length));
+    nonce_ptr++;
+    new_packet_count |= (*nonce_ptr & 0x3FFF) << 16;
+
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+    sl_wfx_host_log("RX packet %lu\n", new_packet_count);
+#endif
+
+    // Update secure link nonce values
+    if (sl_wfx_context->secure_link_nonce.rx_packet_count != new_packet_count) {
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+      sl_wfx_host_log("--SLK counter mismatch-- %lu %lu\r\n", sl_wfx_context->secure_link_nonce.rx_packet_count, new_packet_count);
+#endif
+      result = SL_STATUS_FAIL;
+      SL_WFX_ERROR_CHECK(result);
+    }
+
+    // Encrypted data length is Total bytes read - secure link header -  2 extra bytes read of CTRL register - 2 more bytes for message length in clear
+    uint32_t decrypt_length = read_length - SL_WFX_SECURE_LINK_HEADER_SIZE - SL_WFX_SECURE_LINK_CCM_TAG_SIZE - SL_WFX_CONT_REGISTER_SIZE - 2;
+    result = sl_wfx_host_decode_secure_link_data((uint8_t*)*network_rx_buffer + SL_WFX_SECURE_LINK_HEADER_SIZE + 2,
+                                               decrypt_length,
+                                               sl_wfx_context->secure_link_session_key);
+    SL_WFX_ERROR_CHECK(result);
+
+    sl_wfx_context->secure_link_nonce.rx_packet_count = (sl_wfx_context->secure_link_nonce.rx_packet_count + 1) & ~0xC0000000;
+
+    if ((sl_wfx_context->secure_link_nonce.rx_packet_count > SL_WFX_SECURE_LINK_NONCE_WATERMARK
+         || sl_wfx_context->secure_link_nonce.hp_packet_count > SL_WFX_SECURE_LINK_NONCE_WATERMARK)
+        && sl_wfx_context->secure_link_renegotiation_state == SL_WFX_SECURELINK_DEFAULT) {
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+      sl_wfx_host_log("--SLK renegotiation needed--\r\n");
+#endif
+      sl_wfx_context->secure_link_renegotiation_state = SL_WFX_SECURELINK_RENEGOTIATION_NEEDED;
+    }
+
+    /* Move the buffer pointer by SL_WFX_SECURE_LINK_HEADER_SIZE bytes to point to generic_message_t data */
+    *network_rx_buffer = (sl_wfx_generic_message_t *)((uint8_t *)*network_rx_buffer + SL_WFX_SECURE_LINK_HEADER_SIZE);
+  }
+
+  /* Check received message encryption state corresponds to the expectations from the SLK bitmap */
+  if (sl_wfx_secure_link_encryption_required_get((*network_rx_buffer)->header.id) != has_encrypt_header) {
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+    sl_wfx_host_log("--SLK encryption state mismatch--\r\n");
+#endif
+    result = SL_STATUS_FAIL;
+    SL_WFX_ERROR_CHECK(result);
+  }
+
+  error_handler:
+  return result;
+}
+
+sl_status_t sl_wfx_secure_link_renegotiate(void)
+{
+  sl_status_t result = SL_STATUS_OK;
+
+  if (sl_wfx_context->secure_link_renegotiation_state == SL_WFX_SECURELINK_RENEGOTIATION_NEEDED) {
+#if (SL_WFX_DEBUG_MASK & SL_WFX_DEBUG_SLK)
+    sl_wfx_host_log("--SLK renegotiation pending--\r\n");
+#endif
+    sl_wfx_context->secure_link_renegotiation_state = SL_WFX_SECURELINK_RENEGOTIATION_PENDING;
+    //notify host
+    result = sl_wfx_host_schedule_secure_link_renegotiation();
   }
 
   return result;